Feature/price adjustment

+++ src/main/java/com/BiblioStock/BiblioStock_API/dto/CategoryRequestDTO.java
@@ -1,6 +1,7 @@
 package com.BiblioStock.BiblioStock_API.dto;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record CategoryRequestDTO(
         @NotBlank(message = "O nome da categoria é obrigatório")
@@ -15,5 +16,5 @@ public record CategoryRequestDTO(
         @NotNull(message = "O percentual de reajuste é obrigatório")
         @DecimalMin(value = "0.0", inclusive = false, message = "O percentual deve ser maior que zero")
         @DecimalMax(value = "100.0", message = "O percentual não pode ser maior que 100")
-        Double defaultAdjustmentPercent
+        BigDecimal defaultAdjustmentPercent
 ) {}
